LayoutTests/js/set-foreach-calls-back-with-right-args-expected.txt

Tests to make sure we call forEach callback with right arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS called is true
PASS actual.value is 'value'
PASS actual.key is 'value'
PASS actual.set is s
PASS actual.receiver === receiver is true
PASS successfullyParsed is true

TEST COMPLETE
